Add support for LTO with OS functions
This is achieved by utilizing the 8 bytes of a function stub as a place
to store pointers to the relevant NID and library entry struct. At
runtime these are overwritten by the OS, but this link lets LLVM know
not to remove those seemingly unused sections.

Because this means that the number of function stubs depends on the
linker output, cargo-psp has been modified to properly update the
library stub count after the normal linking stage.

This commit requires linker script changes in rustc; this will happen in
an accompanying patch.